examples/multi_news/requirements.txt
transformers
datasets
sentencepiece
pytorch-ignite

promptsource

einops

wandb
loguru
hydra-core
omegaconf